+++ react/components/form/TextArea.tsx
@@ -68,7 +68,8 @@ const TextArea: React.FunctionComponent<Props> = ({
         placeholder ? formatIOMessage({ id: placeholder, intl }) : ''
       }
       readOnly={readonly}
-      value={value}
+      resize="vertical"
+      value={value ? `${value}` : ''}
     />
   )
 }